Replace Tabs with ResponsiveTabs
